+++ gui/src/components/onboarding/pages/quiz/MoreSetsConfirm.tsx
