+++ gui/src/components/onboarding/pages/quiz/RuntimeQuestion.tsx
@@ -10,7 +10,7 @@ export function QuizRuntimeQuestion() {
   const { applyProgress, setVrcOSC, vrcOsc } = useOnboarding();
   const nav = useNavigate();
 
-  applyProgress(0.4);
+  applyProgress(0.9);
 
   const next = (type: typeof vrcOsc) => {
     setVrcOSC(type);
