+++ gui/src/components/onboarding/pages/quiz/SlimeSetQuestion.tsx
